requirements.txt
# SE445 HW1 – Player Bug Report Categorizer
# Minimal dependencies only

fastapi>=0.115.0
uvicorn>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0

# Google Gemini AI
google-genai>=1.0.0

# Google Sheets via OAuth
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
